Bank Database System
This is a bank database system that users can log in or register as a customer or an employee. Customers can do some transactions like sending money or taking loan. Also, customers can change their account information. Employees can insert money, delete users and approve or deny the loan requests. There is a single security key that allows user to register as an employee.